I need to fix the note detection logic, current implementation is simplistic. Here's how I want it to work:
When user hits record button the first recorded event must be a note (not a pause)
Each event can have variable length but they all need to be normalized. Example: if a note is slighly less than quarter beat it must be recorded as quarter beats.
Maximum length of an event is whole beat.
Minimum length of an event is 16th.
Processing must trigger each 32th
The audio processing needs to detect the dominant tone out of the spectrum, it helps with the accuracy of note detection.